refactor(storage): divide with task, column, history.
